ddrescue/bad_files: add backend base class
This class will be the parent class of all supported backends.
Each backend represents a supported file system and is expected to
implement an __init__ function which receives any binary IO object,
a close() function for resource and context management as well as
a walk() generator function which recursively walks the file system
and yields the full path to the file as well as the numbers of its
initial and final blocks.