bad_files: implement a file class
A file consists of a file system path and extent information
as well as all extents from the GNU ddrescue map which overlap.

This class will act as the point where all the data
from GNU ddrescue map files and supported file systems
is merged so that relevant computations can happen.
Notably, it becomes possible to identify the corrupt parts
of the file, as well as calculate how much corruption there is.